requirements.txt
opencv-python==3.4.15.55
opencv-python-headless==4.5.3.56
opencv-wrapper==0.2.3
json2html==1.3.0
pytesseract==0.3.8
PyYAML==5.3.1
pytest==5.4.1
matplotlib==3.4.2
scikit-image==0.18.2